LayoutTests/storage/indexeddb/database-basics.html

<html>
<head>
<link rel="stylesheet" href="../../fast/js/resources/js-test-style.css">
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="../../fast/js/resources/js-test-post-function.js"></script>
<script src="resources/shared.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description("Test the basics of IndexedDB's IDBDatabase.");
if (window.layoutTestController) 
    layoutTestController.waitUntilDone();

function test()
{
    result = evalAndLog("indexedDB.open('name', 'description')");
    verifyResult(result);
    result.onsuccess = openSuccess;
    result.onerror = unexpectedErrorCallback;
}

function openSuccess()
{
    verifySuccessEvent(event);
    window.db = evalAndLog("db = event.result");

    result = evalAndLog("db.setVersion('new version')");
    verifyResult(result);
    result.onsuccess = setVersionSuccess;
    result.onerror = unexpectedErrorCallback;
}

function setVersionSuccess()
{
    debug("setVersionSuccess():");
    verifySuccessEvent(event);
    window.trans = evalAndLog("trans = event.result");
    shouldBeTrue("trans !== null");
    trans.onabort = unexpectedAbortCallback;

    deleteAllObjectStores(db, testSetVersion);
}

function testSetVersion()
{
    debug("Testing setVersion.");
    result = evalAndLog('db.setVersion("version a")');
    verifyResult(result);
    result.onsuccess = setVersionAgain;
    result.onError = unexpectedErrorCallback;
}

function setVersionAgain()
{
    verifySuccessEvent(event);

    result = evalAndLog('db.setVersion("version b")');
    verifyResult(result);
    result.onsuccess = createObjectStore;
    result.onError = unexpectedErrorCallback;
}

function createObjectStore()
{
    verifySuccessEvent(event);
    shouldBeEqualToString("db.version", "version b");
    shouldBeEqualToString("db.name", "name");
    shouldBe("db.objectStores", "[]");
    shouldBe("db.objectStores.length", "0");
    shouldBe("db.objectStores.contains('')", "false");

    result = evalAndLog('db.createObjectStore("test123")');
    verifyResult(result);
    result.onsuccess = checkObjectStore;
    result.onError = unexpectedErrorCallback;
}

function checkObjectStore()
{
    verifySuccessEvent(event);
    shouldBe("db.objectStores", "['test123']");
    shouldBe("db.objectStores.length", "1");
    shouldBe("db.objectStores.contains('')", "false");
    shouldBe("db.objectStores.contains('test456')", "false");
    shouldBe("db.objectStores.contains('test123')", "true");

    done();
}

test();

var successfullyParsed = true;

</script>
</body>
</html>
